--- GroceryShop.App/AddProduct.cs
@@ -1,44 +1,0 @@
-﻿namespace GroceryShop.App
-{
-    using System;
-    using System.Collections.Generic;
-    using System.ComponentModel;
-    using System.Data;
-    using System.Drawing;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-    using System.Windows.Forms;
-
-    public partial class AddProduct : Form
-    {
-        public AddProduct()
-        {
-            InitializeComponent();
-        }
-
-        //Minimizer button
-        private void btnMinimize_Click(object sender, EventArgs e)
-        {
-            this.WindowState = FormWindowState.Minimized;
-        }
-
-        //close buttion
-        private void btnClose_Click(object sender, EventArgs e)
-        {
-            Application.Exit();
-        }
-        //Mouse hover
-        private void btnClose_MouseEnter(object sender, EventArgs e)
-        {
-            btnClose.ForeColor = Color.Red;
-        }
-        //Mouse hover
-        private void btnClose_MouseLeave(object sender, EventArgs e)
-        {
-            btnClose.ForeColor = Color.White;
-        }
-
-      
-    }
-}
